backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.123.5
h11==0.16.0
httptools==0.7.1
idna==3.11
packaging==25.0
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pytesseract==0.3.13
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
